refac: seperate posts, ui-components plops
